managers/tests/org_api_key_test.py
import pytest
import uuid
from unittest.mock import AsyncMock, patch
from managers import OrgApiKeyManager, OrgApiKey
from models.factory import OrgApiKeyFactory
from sqlalchemy.orm import sessionmaker
from sqlalchemy import create_engine
from models import Base, OrgApiKey
DATABASE_URL = "sqlite:///:memory:"
db_dialect = "sqlite"
class TestOrgApiKeyManager:
    @pytest.fixture(scope="module")
    def engine(self):
        engine = create_engine(DATABASE_URL, echo=True)
        #FKs are not activated by default in sqllite
        with engine.connect() as conn:
            conn.connection.execute("PRAGMA foreign_keys=ON")
        yield engine
        engine.dispose()
    @pytest.fixture
    def session(self, engine):
        Base.metadata.create_all(engine)
        SessionLocal = sessionmaker(bind=engine, expire_on_commit=False)
        session_instance = SessionLocal()
        yield session_instance
        session_instance.close()
    @pytest.fixture
    async def org_api_key_manager(self, session):
        return OrgApiKeyManager(session)
    @pytest.mark.asyncio
    async def test_build(self, org_api_key_manager):
        # Define some mock data for our org_api_key
        mock_data = {
            "name": "Rose",
            "species": "Rosa",
            "age": 2
        }
        # Call the build function of the manager
        org_api_key = await org_api_key_manager.build(**mock_data)
        # Assert that the returned object is an instance of OrgApiKey
        assert isinstance(org_api_key, OrgApiKey)
        # Assert that the attributes of the org_api_key match our mock data
        assert org_api_key.name == mock_data["name"]
        assert org_api_key.species == mock_data["species"]
        assert org_api_key.age == mock_data["age"]
        # Optionally, if the build method has some default values or computations:
        # assert org_api_key.some_attribute == some_expected_value
    @pytest.mark.asyncio
    async def test_build_with_missing_data(self, org_api_key_manager):
        # Define mock data with a missing key
        mock_data = {
            "name": "Rose",
            "age": 2
        }
        # If the build method is expected to raise an exception for missing data, test for that
        with pytest.raises(SomeSpecificException):
            await org_api_key_manager.build(**mock_data)
    @pytest.mark.asyncio
    async def test_add(self, org_api_key_manager, mock_session):
        org_api_key_data = OrgApiKeyFactory.build()
        mock_session.add.return_value = None
        mock_session.commit.return_value = None
        org_api_key = await org_api_key_manager.add(**org_api_key_data)
        mock_session.add.assert_called_once_with(org_api_key)
        mock_session.commit.assert_called_once()
        assert isinstance(org_api_key, OrgApiKey)
    @pytest.mark.asyncio
    async def test_add_correctly_adds_org_api_key_to_database(self, org_api_key_manager, db_session):
        # Create a test org_api_key using the OrgApiKeyFactory without persisting it to the database
        test_org_api_key = OrgApiKeyFactory.build()
        # Add the org_api_key using the manager's add method
        added_org_api_key = await org_api_key_manager.add(org_api_key=test_org_api_key)
        # Fetch the org_api_key from the database directly
        result = await db_session.execute(select(OrgApiKey).filter(OrgApiKey.org_api_key_id == added_org_api_key.org_api_key_id))
        fetched_org_api_key = result.scalars().first()
        # Assert that the fetched org_api_key is not None and matches the added org_api_key
        assert fetched_org_api_key is not None
        assert fetched_org_api_key.org_api_key_id == added_org_api_key.org_api_key_id
        assert fetched_org_api_key.name == added_org_api_key.name
        # ... other attribute checks ...
    @pytest.mark.asyncio
    async def test_add_returns_correct_org_api_key_object(self, org_api_key_manager):
        # Create a test org_api_key using the OrgApiKeyFactory without persisting it to the database
        test_org_api_key = OrgApiKeyFactory.build()
        # Add the org_api_key using the manager's add method
        added_org_api_key = await org_api_key_manager.add(org_api_key=test_org_api_key)
        # Assert that the returned org_api_key matches the test org_api_key
        assert added_org_api_key.org_api_key_id == test_org_api_key.org_api_key_id
        assert added_org_api_key.name == test_org_api_key.name
        # ... other attribute checks ...
    @pytest.mark.asyncio
    async def test_get_by_id(self, org_api_key_manager, mock_session):
        org_api_key_data = OrgApiKeyFactory.build()
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(first=AsyncMock(return_value=org_api_key_data)))
        org_api_key = await org_api_key_manager.get_by_id(1)
        mock_session.execute.assert_called_once()
        assert isinstance(org_api_key, OrgApiKey)
    async def test_get_by_id(self, session: AsyncSession, sample_org_api_key: OrgApiKey):
        manager = OrgApiKeyManager(session)
        retrieved_org_api_key = await manager.get_by_id(sample_org_api_key.org_api_key_id)
        assert retrieved_org_api_key is not None
        assert retrieved_org_api_key.org_api_key_id == sample_org_api_key.org_api_key_id
        assert retrieved_org_api_key.name == "Rose"
        assert retrieved_org_api_key.color == "Red"
    async def test_get_by_id_not_found(self, session: AsyncSession):
        manager = OrgApiKeyManager(session)
        non_existent_id = 9999  # An ID that's not in the database
        retrieved_org_api_key = await manager.get_by_id(non_existent_id)
        assert retrieved_org_api_key is None
    @pytest.mark.asyncio
    async def test_get_by_code_returns_org_api_key(self, org_api_key_manager, db_session):
        # Use your OrgApiKeyFactory to create and save a OrgApiKey object
        code = uuid.uuid4()
        org_api_key = OrgApiKeyFactory(code=code)
        db_session.add(org_api_key)
        await db_session.commit()
        # Fetch the org_api_key using the manager's get_by_code method
        fetched_org_api_key = await org_api_key_manager.get_by_code(code)
        # Assert that the fetched org_api_key is not None and has the expected code
        assert fetched_org_api_key is not None
        assert fetched_org_api_key.code == code
    @pytest.mark.asyncio
    async def test_get_by_code_returns_none_for_nonexistent_code(self, org_api_key_manager):
        # Generate a random UUID that doesn't correspond to any OrgApiKey in the database
        random_code = uuid.uuid4()
        # Try fetching a org_api_key using the manager's get_by_code method
        fetched_org_api_key = await org_api_key_manager.get_by_code(random_code)
        # Assert that the result is None since no org_api_key with the given code exists
        assert fetched_org_api_key is None
    @pytest.mark.asyncio
    async def test_update(self, org_api_key_manager, mock_session):
        org_api_key_data = OrgApiKeyFactory.build()
        updated_data = {"name": "Updated OrgApiKey"}
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(first=AsyncMock(return_value=org_api_key_data)))
        mock_session.commit.return_value = None
        updated_org_api_key = await org_api_key_manager.update(1, **updated_data)
        assert updated_org_api_key.name == "Updated OrgApiKey"
        mock_session.commit.assert_called_once()
        assert isinstance(updated_org_api_key, OrgApiKey)
    async def test_update_valid_org_api_key(self):
        # Mocking a org_api_key instance
        org_api_key = OrgApiKey(org_api_key_id=1, name="Rose", code="ROSE123")
        # Mocking the commit method
        self.session_mock.commit = AsyncMock()
        # Update the org_api_key with new attributes
        updated_org_api_key = await self.manager.update(org_api_key, name="Red Rose", code="REDROSE123")
        # Assertions
        assert updated_org_api_key.name == "Red Rose"
        assert updated_org_api_key.code == "REDROSE123"
        self.session_mock.commit.assert_called_once()
    async def test_update_invalid_org_api_key(self):
        # None org_api_key
        org_api_key = None
        updated_org_api_key = await self.manager.update(org_api_key, name="Red Rose", code="REDROSE123")
        # Assertions
        assert updated_org_api_key is None
        self.session_mock.commit.assert_not_called()
    async def test_update_with_nonexistent_attribute(self):
        # Mocking a org_api_key instance
        org_api_key = OrgApiKey(org_api_key_id=1, name="Rose", code="ROSE123")
        # Mocking the commit method
        self.session_mock.commit = AsyncMock()
        # This should raise an AttributeError since 'color' is not an attribute of OrgApiKey
        with pytest.raises(AttributeError):
            await self.manager.update(org_api_key, color="Red")
        self.session_mock.commit.assert_not_called()
    @pytest.mark.asyncio
    async def test_delete(self, org_api_key_manager, mock_session):
        org_api_key_data = OrgApiKeyFactory.build()
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(first=AsyncMock(return_value=org_api_key_data)))
        mock_session.delete.return_value = None
        mock_session.commit.return_value = None
        deleted_org_api_key = await org_api_key_manager.delete(1)
        mock_session.delete.assert_called_once_with(deleted_org_api_key)
        mock_session.commit.assert_called_once()
        assert isinstance(deleted_org_api_key, OrgApiKey)
    @pytest.mark.asyncio
    async def test_delete_nonexistent(self, org_api_key_manager, mock_session):
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(first=AsyncMock(return_value=None)))
        with pytest.raises(ValueError, match="OrgApiKey not found"):
            await org_api_key_manager.delete(999)
    @pytest.mark.asyncio
    async def test_get_list(self, org_api_key_manager, mock_session):
        org_api_keys_data = [OrgApiKeyFactory.build() for _ in range(5)]
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(all=AsyncMock(return_value=org_api_keys_data)))
        org_api_keys = await org_api_key_manager.get_list()
        mock_session.execute.assert_called_once()
        assert len(org_api_keys) == 5
        assert all(isinstance(org_api_key, OrgApiKey) for org_api_key in org_api_keys)
    @pytest.mark.asyncio
    async def test_to_json(self, org_api_key_manager):
        org_api_key_data = OrgApiKeyFactory.build()
        org_api_key = OrgApiKey(**org_api_key_data)
        json_data = org_api_key_manager.to_json(org_api_key)
        assert json_data is not None
        # You might want to do more specific checks on the JSON structure
    @pytest.mark.asyncio
    async def test_from_json(self, org_api_key_manager):
        org_api_key_data = OrgApiKeyFactory.build()
        org_api_key = OrgApiKey(**org_api_key_data)
        json_data = org_api_key_manager.to_json(org_api_key)
        deserialized_org_api_key = org_api_key_manager.from_json(json_data)
        assert isinstance(deserialized_org_api_key, OrgApiKey)
        # Additional checks on the deserialized data can be added
    @pytest.mark.asyncio
    async def test_add_bulk(self, org_api_key_manager, mock_session):
        org_api_keys_data = [OrgApiKeyFactory.build() for _ in range(5)]
        mock_session.add_all.return_value = None
        mock_session.commit.return_value = None
        org_api_keys = await org_api_key_manager.add_bulk(org_api_keys_data)
        mock_session.add_all.assert_called_once()
        mock_session.commit.assert_called_once()
        assert len(org_api_keys) == 5
    @pytest.mark.asyncio
    async def test_update_bulk_success():
        manager = OrgApiKeyManager()
        session_mock = AsyncMock()
        manager.session = session_mock
        # Mocking org_api_key instances
        org_api_key1 = OrgApiKey(org_api_key_id=1, name="Rose", code="ROSE123")
        org_api_key2 = OrgApiKey(org_api_key_id=2, name="Tulip", code="TULIP123")
        # Mocking the get_by_id method to return the corresponding org_api_key
        async def mock_get_by_id(org_api_key_id):
            if org_api_key_id == 1:
                return org_api_key1
            if org_api_key_id == 2:
                return org_api_key2
        manager.get_by_id = mock_get_by_id
        # Mocking the commit method
        session_mock.commit = AsyncMock()
        # Update org_api_keys
        updates = [{"org_api_key_id": 1, "name": "Red Rose"}, {"org_api_key_id": 2, "name": "Yellow Tulip"}]
        updated_org_api_keys = await manager.update_bulk(updates)
        # Assertions
        assert len(updated_org_api_keys) == 2
        assert updated_org_api_keys[0].name == "Red Rose"
        assert updated_org_api_keys[1].name == "Yellow Tulip"
        session_mock.commit.assert_called_once()
    @pytest.mark.asyncio
    async def test_update_bulk_missing_org_api_key_id():
        manager = OrgApiKeyManager()
        # No org_api_keys to update since org_api_key_id is missing
        updates = [{"name": "Red Rose"}]
        updated_org_api_keys = await manager.update_bulk(updates)
        # Assertions
        assert len(updated_org_api_keys) == 0
    @pytest.mark.asyncio
    async def test_update_bulk_org_api_key_not_found():
        manager = OrgApiKeyManager()
        session_mock = AsyncMock()
        manager.session = session_mock
        # Mocking the get_by_id method to return None (org_api_key not found)
        manager.get_by_id = AsyncMock(return_value=None)
        # Mocking the commit method
        session_mock.commit = AsyncMock()
        # Update org_api_keys
        updates = [{"org_api_key_id": 1, "name": "Red Rose"}]
        updated_org_api_keys = await manager.update_bulk(updates)
        # Assertions
        assert len(updated_org_api_keys) == 0
        session_mock.commit.assert_not_called()
    @pytest.mark.asyncio
    async def test_delete_bulk_success():
        manager = OrgApiKeyManager()
        session_mock = AsyncMock()
        manager.session = session_mock
        # Mocking org_api_key instances
        org_api_key1 = OrgApiKey(org_api_key_id=1, name="Rose", code="ROSE123")
        org_api_key2 = OrgApiKey(org_api_key_id=2, name="Tulip", code="TULIP123")
        # Mocking the get_by_id method to return the corresponding org_api_key
        async def mock_get_by_id(org_api_key_id):
            if org_api_key_id == 1:
                return org_api_key1
            if org_api_key_id == 2:
                return org_api_key2
        manager.get_by_id = mock_get_by_id
        # Mocking the commit and delete methods
        session_mock.commit = AsyncMock()
        session_mock.delete = AsyncMock()
        # Delete org_api_keys
        org_api_key_ids = [1, 2]
        result = await manager.delete_bulk(org_api_key_ids)
        # Assertions
        assert result is True
        session_mock.delete.assert_called()
        session_mock.commit.assert_called_once()
    @pytest.mark.asyncio
    async def test_delete_bulk_some_org_api_keys_not_found():
        manager = OrgApiKeyManager()
        session_mock = AsyncMock()
        manager.session = session_mock
        # Mocking the get_by_id method to return None (org_api_key not found)
        async def mock_get_by_id(org_api_key_id):
            if org_api_key_id == 1:
                return None
            if org_api_key_id == 2:
                return OrgApiKey(org_api_key_id=2, name="Tulip", code="TULIP123")
        manager.get_by_id = mock_get_by_id
        # Mocking the commit and delete methods
        session_mock.commit = AsyncMock()
        session_mock.delete = AsyncMock()
        # Delete org_api_keys
        org_api_key_ids = [1, 2]
        result = await manager.delete_bulk(org_api_key_ids)
        # Assertions
        assert result is True
        session_mock.delete.assert_called_once_with(OrgApiKey(org_api_key_id=2, name="Tulip", code="TULIP123"))
        session_mock.commit.assert_called_once()
    @pytest.mark.asyncio
    async def test_delete_bulk_empty_list():
        manager = OrgApiKeyManager()
        session_mock = AsyncMock()
        manager.session = session_mock
        # Mocking the commit and delete methods
        session_mock.commit = AsyncMock()
        session_mock.delete = AsyncMock()
        # Delete org_api_keys with an empty list
        org_api_key_ids = []
        result = await manager.delete_bulk(org_api_key_ids)
        # Assertions
        assert result is True
        session_mock.delete.assert_not_called()
        session_mock.commit.assert_not_called()
    @pytest.mark.asyncio
    async def test_count(self, org_api_key_manager, mock_session):
        org_api_keys_data = [OrgApiKeyFactory.build() for _ in range(5)]
        mock_session.execute.return_value = AsyncMock(scalars=AsyncMock(all=AsyncMock(return_value=org_api_keys_data)))
        count = await org_api_key_manager.count()
        mock_session.execute.assert_called_once()
        assert count == 5
    @pytest.mark.asyncio
    async def test_count_basic_functionality(async_session):
        # Add a org_api_key
        new_org_api_key = OrgApiKey()
        async_session.add(new_org_api_key)
        await async_session.commit()
        manager = YourManagerClass(session=async_session)
        count = await manager.count()
        assert count == 1
    @pytest.mark.asyncio
    async def test_count_empty_database(async_session):
        manager = YourManagerClass(session=async_session)
        count = await manager.count()
        assert count == 0
    @pytest.mark.asyncio
    async def test_count_multiple_additions(async_session):
        # Add multiple org_api_keys
        org_api_keys = [OrgApiKey() for _ in range(5)]
        async_session.add_all(org_api_keys)
        await async_session.commit()
        manager = YourManagerClass(session=async_session)
        count = await manager.count()
        assert count == 5
    @pytest.mark.asyncio
    async def test_count_database_connection_issues(async_session, mocker):
        # Mock the session's execute method to simulate a database connection error
        mocker.patch.object(async_session, 'execute', side_effect=Exception("DB connection error"))
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.count()
    @pytest.mark.asyncio
    async def test_get_sorted_list_basic_sorting(async_session):
        # Add org_api_keys
        org_api_keys = [OrgApiKey(name=f"OrgApiKey_{i}") for i in range(5)]
        async_session.add_all(org_api_keys)
        await async_session.commit()
        manager = YourManagerClass(session=async_session)
        sorted_org_api_keys = await manager.get_sorted_list(sort_by="name")
        assert [org_api_key.name for org_api_key in sorted_org_api_keys] == [f"OrgApiKey_{i}" for i in range(5)]
    @pytest.mark.asyncio
    async def test_get_sorted_list_descending_sorting(async_session):
        # Add org_api_keys
        org_api_keys = [OrgApiKey(name=f"OrgApiKey_{i}") for i in range(5)]
        async_session.add_all(org_api_keys)
        await async_session.commit()
        manager = YourManagerClass(session=async_session)
        sorted_org_api_keys = await manager.get_sorted_list(sort_by="name", order="desc")
        assert [org_api_key.name for org_api_key in sorted_org_api_keys] == [f"OrgApiKey_{i}" for i in reversed(range(5))]
    @pytest.mark.asyncio
    async def test_get_sorted_list_invalid_attribute(async_session):
        manager = YourManagerClass(session=async_session)
        with pytest.raises(AttributeError):
            await manager.get_sorted_list(sort_by="invalid_attribute")
    @pytest.mark.asyncio
    async def test_get_sorted_list_database_connection_issues(async_session, mocker):
        # Mock the session's execute method to simulate a database connection error
        mocker.patch.object(async_session, 'execute', side_effect=Exception("DB connection error"))
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.get_sorted_list(sort_by="name")
    @pytest.mark.asyncio
    async def test_get_sorted_list_empty_database(async_session):
        manager = YourManagerClass(session=async_session)
        sorted_org_api_keys = await manager.get_sorted_list(sort_by="name")
        assert len(sorted_org_api_keys) == 0
    @pytest.mark.asyncio
    async def test_refresh_basic(async_session):
        # Add a org_api_key
        org_api_key = OrgApiKey(name="OrgApiKey_1")
        async_session.add(org_api_key)
        await async_session.commit()
        # Modify the org_api_key directly in the database
        await async_session.execute('UPDATE org_api_keys SET name = :new_name WHERE id = :org_api_key_id', {"new_name": "Modified_OrgApiKey", "org_api_key_id": org_api_key.id})
        await async_session.commit()
        # Now, refresh the org_api_key using the manager function
        manager = YourManagerClass(session=async_session)
        refreshed_org_api_key = await manager.refresh(org_api_key)
        assert refreshed_org_api_key.name == "Modified_OrgApiKey"
    @pytest.mark.asyncio
    async def test_refresh_nonexistent_org_api_key(async_session):
        org_api_key = OrgApiKey(id=999, name="Nonexistent_OrgApiKey")
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception):  # Modify the exception type based on your ORM's behavior
            await manager.refresh(org_api_key)
    @pytest.mark.asyncio
    async def test_refresh_database_connection_issues(async_session, mocker):
        # Mock the session's refresh method to simulate a database connection error
        mocker.patch.object(async_session, 'refresh', side_effect=Exception("DB connection error"))
        org_api_key = OrgApiKey(name="OrgApiKey_1")
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.refresh(org_api_key)
    @pytest.mark.asyncio
    async def test_exists_with_existing_org_api_key(async_session):
        # Add a org_api_key
        org_api_key = OrgApiKey(name="OrgApiKey_1")
        async_session.add(org_api_key)
        await async_session.commit()
        # Check if the org_api_key exists using the manager function
        manager = YourManagerClass(session=async_session)
        assert await manager.exists(org_api_key.id) == True
    @pytest.mark.asyncio
    async def test_exists_with_nonexistent_org_api_key(async_session):
        non_existent_id = 999
        manager = YourManagerClass(session=async_session)
        assert await manager.exists(non_existent_id) == False
    @pytest.mark.asyncio
    async def test_exists_with_invalid_id_type(async_session):
        invalid_id = "invalid_id"
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception):  # Modify the exception type based on your ORM's behavior or validation
            await manager.exists(invalid_id)
    @pytest.mark.asyncio
    async def test_exists_database_connection_issues(async_session, mocker):
        # Mock the get_by_id method to simulate a database connection error
        mocker.patch.object(YourManagerClass, 'get_by_id', side_effect=Exception("DB connection error"))
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.exists(1)
    #get_by_organization_id
    #get_by_org_customer_id
    @pytest.mark.asyncio
    async def test_get_by_org_customer_id_existing(async_session):
        # Add a org_api_key with a specific org_customer_id
        org_api_key = OrgApiKey(name="OrgApiKey_1", org_customer_id=5)
        async_session.add(org_api_key)
        await async_session.commit()
        # Fetch the org_api_key using the manager function
        manager = YourManagerClass(session=async_session)
        fetched_org_api_keys = await manager.get_by_org_customer_id(5)
        assert len(fetched_org_api_keys) == 1
        assert fetched_org_api_keys[0].name == "OrgApiKey_1"
    @pytest.mark.asyncio
    async def test_get_by_org_customer_id_nonexistent(async_session):
        non_existent_id = 999
        manager = YourManagerClass(session=async_session)
        fetched_org_api_keys = await manager.get_by_org_customer_id(non_existent_id)
        assert len(fetched_org_api_keys) == 0
    @pytest.mark.asyncio
    async def test_get_by_org_customer_id_invalid_type(async_session):
        invalid_id = "invalid_id"
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception):  # Modify the exception type based on your ORM's behavior or validation
            await manager.get_by_org_customer_id(invalid_id)
    @pytest.mark.asyncio
    async def test_get_by_org_customer_id_database_connection_issues(async_session, mocker):
        # Mock the execute method to simulate a database connection error
        mocker.patch.object(async_session, 'execute', side_effect=Exception("DB connection error"))
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.get_by_org_customer_id(1)
    @pytest.mark.asyncio
    async def test_get_by_organization_id_existing(async_session):
        # Add a org_api_key with a specific organization_id
        org_api_key = OrgApiKey(name="OrgApiKey_1", organization_id=5)
        async_session.add(org_api_key)
        await async_session.commit()
        # Fetch the org_api_key using the manager function
        manager = YourManagerClass(session=async_session)
        fetched_org_api_keys = await manager.get_by_organization_id(5)
        assert len(fetched_org_api_keys) == 1
        assert fetched_org_api_keys[0].name == "OrgApiKey_1"
    @pytest.mark.asyncio
    async def test_get_by_organization_id_nonexistent(async_session):
        non_existent_id = 999
        manager = YourManagerClass(session=async_session)
        fetched_org_api_keys = await manager.get_by_organization_id(non_existent_id)
        assert len(fetched_org_api_keys) == 0
    @pytest.mark.asyncio
    async def test_get_by_organization_id_invalid_type(async_session):
        invalid_id = "invalid_id"
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception):  # Modify the exception type based on your ORM's behavior or validation
            await manager.get_by_organization_id(invalid_id)
    @pytest.mark.asyncio
    async def test_get_by_organization_id_database_connection_issues(async_session, mocker):
        # Mock the execute method to simulate a database connection error
        mocker.patch.object(async_session, 'execute', side_effect=Exception("DB connection error"))
        manager = YourManagerClass(session=async_session)
        with pytest.raises(Exception, match="DB connection error"):
            await manager.get_by_organization_id(1)
